Break out deposits code from server
