+++ lib/ui/src/Cardano/Wallet/UI/Deposit/Server.hs
@@ -9,11 +9,6 @@ module Cardano.Wallet.UI.Deposit.Server
     ( serveUI
     ) where
 
-import Prelude
-
-import Cardano.Slotting.Slot
-    ( WithOrigin (..)
-    )
 import Cardano.Wallet.Api.Http.Server.Handlers.NetworkInformation
     ( getNetworkInformation
     )
@@ -23,21 +18,11 @@ import Cardano.Wallet.Api.Types
 import Cardano.Wallet.Deposit.IO
     ( WalletBootEnv
     )
-import Cardano.Wallet.Deposit.IO.Network.Mock
-    ( originTime
-    )
 import Cardano.Wallet.Deposit.IO.Network.Type
     ( NetworkEnv
     )
-import Cardano.Wallet.Deposit.Map
-    ( unPatch
-    )
-import Cardano.Wallet.Deposit.Pure
-    ( Customer
-    )
 import Cardano.Wallet.Deposit.REST
     ( WalletResource
-    , WalletResourceM
     , deleteWallet
     , initXPubWallet
     )
@@ -48,9 +33,6 @@ import Cardano.Wallet.Primitive.NetworkId
     ( HasSNetworkId (..)
     , SNetworkId
     , networkIdVal
-    )
-import Cardano.Wallet.Read
-    ( TxId
     )
 import Cardano.Wallet.Shelley.BlockchainSource
     ( BlockchainSource (..)
@@ -94,10 +76,6 @@ import Cardano.Wallet.UI.Common.Html.Pages.Template.Head
 import Cardano.Wallet.UI.Common.Html.Pages.Wallet
     ( mnemonicH
     )
-import Cardano.Wallet.UI.Common.Html.Scrolling
-    ( Scrolling (..)
-    , newScrolling
-    )
 import Cardano.Wallet.UI.Common.Layer
     ( SessionLayer (..)
     , UILayer (..)
@@ -108,10 +86,7 @@ import Cardano.Wallet.UI.Cookies
     , sessioning
     )
 import Cardano.Wallet.UI.Deposit.API
-    ( DepositsParams (..)
-    , DownTime
-    , Expand
-    , TransactionHistoryParams
+    ( TransactionHistoryParams
     , UI
     , settingsSseToggleLink
     )
@@ -122,21 +97,8 @@ import Cardano.Wallet.UI.Deposit.Handlers.Addresses
 import Cardano.Wallet.UI.Deposit.Handlers.Addresses.Transactions
     ( getCustomerHistory
     )
-import Cardano.Wallet.UI.Deposit.Handlers.Deposits
-    ( depositCustomersHandler
-    , depositCustomersPaginationHandlers
-    , depositsPaginationHandlers
-    )
-import Cardano.Wallet.UI.Deposit.Handlers.Deposits.Fake
-    ( getFakeDepositsHistory
-    )
-import Cardano.Wallet.UI.Deposit.Handlers.Deposits.TxIds
-    ( depositCustomersTxIdsHandler
-    , depositCustomersTxIdsPaginationHandlers
-    )
 import Cardano.Wallet.UI.Deposit.Handlers.Lib
-    ( catchRunWalletResourceM
-    , walletPresence
+    ( walletPresence
     )
 import Cardano.Wallet.UI.Deposit.Handlers.Wallet
     ( deleteWalletHandler
@@ -152,16 +114,6 @@ import Cardano.Wallet.UI.Deposit.Html.Pages.Addresses.Transactions
     ( customerHistoryH
     , transactionsElementH
     )
-import Cardano.Wallet.UI.Deposit.Html.Pages.Deposits
-    ( depositByCustomerH
-    , depositH
-    , depositsElementH
-    , scrollableDeposits
-    , scrollableDepositsCustomers
-    )
-import Cardano.Wallet.UI.Deposit.Html.Pages.Deposits.TxIds
-    ( scrollableDepositsCustomersTxIds
-    )
 import Cardano.Wallet.UI.Deposit.Html.Pages.Page
     ( Page (..)
     , headerElementH
@@ -171,48 +123,40 @@ import Cardano.Wallet.UI.Deposit.Html.Pages.Wallet
     ( deleteWalletModalH
     , walletElementH
     )
+import Cardano.Wallet.UI.Deposit.Server.Deposits
+    ( serveDeposits
+    , serveDepositsCustomersTxIds
+    , serveDepositsCustomersTxIdsPage
+    , serveDepositsHistory
+    , serveDepositsHistoryPage
+    , serveDepositsHistoryWindow
+    , serveDepositsHistoryWindowPage
+    )
+import Cardano.Wallet.UI.Deposit.Server.Lib
+    ( alert
+    , origin
+    , renderSmoothHtml
+    , showTime
+    )
 import Control.Monad.Trans
     ( MonadIO (..)
     )
 import Control.Tracer
     ( Tracer (..)
     )
-import Data.Bifunctor
-    ( first
-    )
-import Data.Foldable
-    ( fold
-    )
 import Data.Functor
     ( ($>)
-    )
-import Data.Monoid
-    ( First (..)
-    )
-import Data.Ord
-    ( Down (..)
     )
 import Data.Text
     ( Text
     )
 import Data.Time
-    ( UTCTime
-    , defaultTimeLocale
-    , formatTime
-    , getCurrentTime
-    )
-import Data.Time.Clock.POSIX
-    ( posixSecondsToUTCTime
-    )
-import Lucid
-    ( Html
-    , ToHtml (..)
-    , class_
-    , div_
+    ( getCurrentTime
     )
 import Paths_cardano_wallet_ui
     ( getDataFileName
     )
+import Prelude
 import Servant
     ( Handler
     , Server
@@ -221,9 +165,6 @@ import Servant
 
 import qualified Cardano.Read.Ledger.Block.Block as Read
 import qualified Data.ByteString.Lazy as BL
-
-showTime :: UTCTime -> String
-showTime = formatTime defaultTimeLocale "%Y-%m-%d %H:%M:%S"
 
 serveUI
     :: forall n x
@@ -296,9 +237,6 @@ serveUI tr network ul env dbDir config _ nl bs =
     wsl f = withSessionLayer ul $ \l -> f l
     initWallet = initXPubWallet tr env dbDir
 
-alert :: ToHtml a => a -> RawHtml
-alert = renderHtml . alertH
-
 serveFakeDataBackground :: Handler BL.ByteString
 serveFakeDataBackground = do
     file <- liftIO $ getDataFileName "data/images/fake-data.png"
@@ -320,12 +258,6 @@ serveCustomerHistory network ul params = do
                 alertH
                 params
 
-renderSmoothHtml :: Html () -> RawHtml
-renderSmoothHtml response =
-    renderHtml
-        $ div_ [class_ "smooth"]
-        $ toHtml response
-
 serveFavicon :: Handler BL.ByteString
 serveFavicon = do
     file <- liftIO $ getDataFileName "data/images/icon.png"
@@ -342,191 +274,3 @@ serveTransactions ul =
             pure
                 $ renderSmoothHtml
                 $ transactionsElementH now origin
-
-origin :: UTCTime
-origin = posixSecondsToUTCTime $ fromIntegral originTime
-
-serveDeposits
-    :: UILayer WalletResource
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDeposits ul = withSessionLayer ul $ \layer -> do
-    wp <- walletPresence layer
-    pure
-        $ renderSmoothHtml
-        $ depositsElementH alertH wp
-
-depositsTable
-    :: DepositsParams
-    -> WalletResourceM (Scrolling WalletResourceM DownTime)
-depositsTable params = do
-    let hs = depositsPaginationHandlers params getFakeDepositsHistory 100
-    newScrolling $ scrollableDeposits params hs
-
-serveDepositsHistoryPage
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> Maybe (WithOrigin UTCTime)
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsHistoryPage ul params (Just index) = withSessionLayer ul
-    $ \layer -> do
-        result <- catchRunWalletResourceM layer $ do
-            scrolling <- depositsTable params
-            scroll scrolling (depositsPages params) $ Down index
-        pure $ renderHtml result
-serveDepositsHistoryPage ul _ _ = withSessionLayer ul
-    $ \_layer -> do
-        pure
-            $ renderHtml
-            $ alertH ("No page index provided" :: Text)
-
-depositsCustomersTable
-    :: DepositsParams
-    -> DownTime
-    -> WalletResourceM (Scrolling WalletResourceM Customer)
-depositsCustomersTable params time = do
-    let hs =
-            depositCustomersPaginationHandlers
-                params
-                getFakeDepositsHistory
-                time
-                100
-    newScrolling $ scrollableDepositsCustomers params time hs
-
-depositsCustomersTxIdsTable
-    :: DepositsParams
-    -> DownTime
-    -> Customer
-    -> WalletResourceM (Scrolling WalletResourceM TxId)
-depositsCustomersTxIdsTable params time customer = do
-    let hs =
-            depositCustomersTxIdsPaginationHandlers
-                params
-                getFakeDepositsHistory
-                time
-                customer
-                100
-    newScrolling $ scrollableDepositsCustomersTxIds params time customer hs
-
-serveDepositsHistoryWindowPage
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> (Maybe (WithOrigin UTCTime))
-    -> Maybe Customer
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsHistoryWindowPage ul params (Just time) (Just customer) =
-    withSessionLayer ul
-        $ \layer -> do
-            result <- catchRunWalletResourceM layer $ do
-                scrolling <- depositsCustomersTable params (Down time)
-                scroll scrolling (depositsCustomersPages params) customer
-            pure $ renderHtml result
-serveDepositsHistoryWindowPage ul _ _ _ = withSessionLayer ul
-    $ \_layer -> do
-        pure
-            $ renderHtml
-            $ alertH ("No time or customer provided" :: Text)
-
-serveDepositsHistory
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsHistory ul params = withSessionLayer ul $ \layer -> do
-    result <- catchRunWalletResourceM layer $ do
-        scrolling <- depositsTable params
-        pure $ widget scrolling []
-    pure $ renderSmoothHtml result
-
-serveDepositsHistoryWindow
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> Maybe (WithOrigin UTCTime)
-    -> Maybe Expand
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsHistoryWindow ul params mtime mexpand = withSessionLayer ul
-    $ \layer -> do
-        fmap renderHtml $ case mtime of
-            Nothing -> pure $ alertH ("No time provided" :: Text)
-            Just time -> do
-                result <- catchRunWalletResourceM layer $ do
-                    scrolling <- depositsCustomersTable params (Down time)
-                    pure $ widget scrolling
-                depositCustomersHandler
-                    layer
-                    ( \window ->
-                        depositH
-                            params
-                            mexpand
-                            ( Down time
-                            , first getFirst
-                                $ fold
-                                $ unPatch window
-                            )
-                            result
-                    )
-                    alertH
-                    params
-                    time
-
-serveDepositsCustomersTxIds
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> Maybe (WithOrigin UTCTime)
-    -> Maybe Customer
-    -> Maybe Expand
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsCustomersTxIds ul params (Just time) (Just customer) mexpand =
-    withSessionLayer ul $ \layer -> do
-        fmap renderHtml $ do
-            result <- catchRunWalletResourceM layer $ do
-                scrolling <- depositsCustomersTxIdsTable params (Down time) customer
-                pure $ widget scrolling
-            depositCustomersTxIdsHandler
-                layer
-                ( \txIds ->
-                    depositByCustomerH
-                        params
-                        mexpand
-                        (Down time)
-                        ( customer
-                        , first getFirst
-                            $ fold
-                            $ unPatch txIds
-                        )
-                        result
-                )
-                alertH
-                params
-                time
-                customer
-serveDepositsCustomersTxIds ul _ _ _ _ = withSessionLayer ul
-    $ \_layer -> do
-        pure
-            $ renderHtml
-            $ alertH ("No time or customer provided" :: Text)
-
-serveDepositsCustomersTxIdsPage
-    :: UILayer WalletResource
-    -> DepositsParams
-    -> Maybe (WithOrigin UTCTime)
-    -> Maybe Customer
-    -> Maybe TxId
-    -> Maybe RequestCookies
-    -> Handler (CookieResponse RawHtml)
-serveDepositsCustomersTxIdsPage ul params (Just time) (Just customer) (Just txId) =
-    withSessionLayer ul
-        $ \layer -> do
-            result <- catchRunWalletResourceM layer $ do
-                scrolling <- depositsCustomersTxIdsTable params (Down time) customer
-                scroll scrolling (depositsCustomersTxIdsPages params) txId
-            pure $ renderHtml result
-serveDepositsCustomersTxIdsPage ul _ _ _ _ = withSessionLayer ul
-    $ \_layer -> do
-        pure
-            $ renderHtml
-            $ alertH ("No time, customer or txId provided" :: Text)
